[feat] merge fix (#64)
